(feat) AntiVuln lib separated into separate module for consumption outside of the testing monorepo.
